- New structure for the tests created. - Added tests for every endpoint for the customers. Currently we are testing only the controllers. - Added JsonContent property in the RowJsonActionResult. - Created custom fields validator. - Code improvements.

+++ src/Plugins/Nop.Plugin.Api/Models/CustomersParameters/CustomersSearchParametersModel.cs
@@ -19,7 +19,7 @@ namespace Nop.Plugin.Api.Models.CustomersParameters
         public string Order { get; set; }
         public string Query { get; set; }
         public int Page { get; set; }
-        public byte Limit { get; set; }
+        public int Limit { get; set; }
         public string Fields { get; set; }
     }
 }
